backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.35
PyJWT==2.8.0
bcrypt==4.1.2
pytest==7.4.3
pytest-flask==1.3.0